refactor meta data to cache class&remove token NotImplemented
