+++ kvpress/ada_cache.py
@@ -3,10 +3,83 @@
 # This file is part of the YuanFeng project and is licensed under the MIT License.
 # SPDX-License-Identifier: MIT
 
+from pickle import LIST
+from attr import dataclass
 from transformers.cache_utils import Cache
 from typing import List, Optional, Tuple
 import torch
 
+
+@dataclass
+class MetaData:
+    decoding_cu_seqlens_q: torch.Tensor = None
+    cu_seqlens_k: torch.Tensor = None
+    max_seqlen_k: int = None
+    cu_offset: torch.Tensor = None
+    cu_head_offset: torch.Tensor = None
+    head_lens: torch.Tensor = None 
+    bsz: int = None 
+    num_key_value_heads: int = None
+    seen_tokens: int = None
+
+    def _update_metadata_while_compressing(self, head_lens, cu_seqlens_k,max_seqlen_k):
+        self.head_lens = head_lens
+        self.cu_seqlens_k = cu_seqlens_k
+        self.max_seqlen_k = max_seqlen_k
+
+    def _update_metadata_remove_n(self, n):
+        self.max_seqlen_k -= n
+        self.seen_tokens -= n
+        self.head_lens -= n
+        self.cu_seqlens_k -= self.cu_offset * n
+
+    def _update_metadata(self, key_states):
+        bs, head, seqlen, dim = key_states.shape
+
+        self.max_seqlen_k += seqlen
+        self.cu_seqlens_k += self.cu_offset * seqlen
+        self.head_lens += seqlen
+        self.seen_tokens += seqlen
+
+    # init the metadata for the flattened cache during the prefilling phase
+    def _init_metadata(self, key_states):
+
+        """
+        this method is used to initialize metadata for the flatten cache,
+        input key_states is a regular key states with shape [bsz, num_key_value_heads, seqlen, head_dim]
+        """
+
+        bsz, num_key_value_heads, k_len, head_dim = key_states.shape
+        k_lens = bsz *  num_key_value_heads * [k_len]
+        _device = key_states.device
+        max_seqlen_k = max(k_lens)
+
+        head_seqlens_k = torch.tensor(k_lens, dtype=torch.int32, device=_device)
+        cu_seqlens = torch.cumsum(head_seqlens_k, dim=0, dtype=torch.int32)
+        cu_seqlens_k = torch.cat(
+            [torch.tensor([0], dtype=torch.int32, device=_device), cu_seqlens], dim=0)
+        
+
+        decoding_q_lens = bsz *  num_key_value_heads * [1] 
+        decoding_head_seqlens_q = torch.tensor(decoding_q_lens, dtype=torch.int32,device=_device)
+        decoding_cu_seqlens_q = torch.cumsum(decoding_head_seqlens_q, dim=0, dtype=torch.int32)
+        decoding_cu_seqlens_q = torch.cat(
+            [ torch.tensor([0], dtype=torch.int32, device=_device), decoding_cu_seqlens_q], dim=0)
+        
+        
+        cu_offset = torch.arange(0, bsz * num_key_value_heads + 1, dtype=torch.int32, device=_device)
+        cu_head_offset = torch.arange(1, bsz * num_key_value_heads + 1, dtype=torch.int32, device=_device)
+
+        # init metadata
+        self.decoding_cu_seqlens_q = decoding_cu_seqlens_q
+        self.cu_seqlens_k = cu_seqlens_k
+        self.max_seqlen_k = max_seqlen_k
+        self.cu_offset = cu_offset
+        self.cu_head_offset = cu_head_offset
+        self.head_lens = head_seqlens_k
+        self.bsz = bsz
+        self.num_key_value_heads = num_key_value_heads
+        self.seen_tokens = k_len
 
 class DynamicCacheSplitHeadFlatten(Cache):
     
@@ -19,6 +92,7 @@ class DynamicCacheSplitHeadFlatten(Cache):
         self.key_cache: List[List[torch.Tensor]] = []
         self.value_cache: List[List[torch.Tensor]] = []
         self._seen_tokens = 0
+        self.metadata_list:List[MetaData] = []
 
     def __len__(self):
         return len(self.key_cache)
@@ -37,7 +111,7 @@ class DynamicCacheSplitHeadFlatten(Cache):
         # NOTE: k, v = [head_num]( bs, 1, seqlen, dim)
         # each layer is a flatten layout like:
         # [bsz * (head_0_len + head_1_len + ...+ head_n_len) , dim]
-        attn = cache_kwargs.get("attn", None)
+        # attn = cache_kwargs.get("attn", None)
         if len(self.key_cache) <= layer_idx:
             # prefilling
             # flatten key and value
@@ -46,10 +120,12 @@ class DynamicCacheSplitHeadFlatten(Cache):
             flatten_value_cache = value_states.reshape(bs* head_num* seqlen, head_dim)
             self.key_cache.append(flatten_key_cachee)
             self.value_cache.append(flatten_value_cache)
-
+            meta_data = MetaData()
+            meta_data._init_metadata(key_states)
+            self.metadata_list.append(meta_data)
             # init metadata for flatten key states
-            attn.metadata._init_metadata(key_states)
-            self._seen_tokens = attn.metadata.seen_tokens
+            # self.metadata._init_metadata(key_states)
+            self._seen_tokens = seqlen
         else:
             # decoding
             assert self.key_cache[layer_idx].dim() == 2
@@ -58,8 +134,8 @@ class DynamicCacheSplitHeadFlatten(Cache):
             # TODO: Currently only support bs == 1
             assert bs == 1 , f"bs: {bs}"
             # NOTE: phase 2. we got [bs, head, seqlen, dim] as k, v input
-            head_lens = attn.metadata.head_lens
-            cu_seqlens_k = attn.metadata.cu_seqlens_k
+            head_lens = self.metadata_list[layer_idx].head_lens
+            cu_seqlens_k = self.metadata_list[layer_idx].cu_seqlens_k
 
             # TODO: wrap as a python interface
             from tiny_api_cuda import update_flatten_klenN_view
@@ -70,9 +146,8 @@ class DynamicCacheSplitHeadFlatten(Cache):
             self.value_cache[layer_idx] = new_value_cache
 
             # update metadata
-            attn.metadata._update_metadata(key_states)
-            self._seen_tokens = attn.metadata.seen_tokens
-
+            self.metadata_list[layer_idx]._update_metadata(key_states)
+            self._seen_tokens += seqlen
         return self.key_cache[layer_idx], self.value_cache[layer_idx]
 
     def get_seq_length(self, layer_idx: Optional[int] = 0) -> int:
@@ -81,6 +156,16 @@ class DynamicCacheSplitHeadFlatten(Cache):
 
         # TODO: return 1 to means has content for now
         return 1
+
+    def remove_tokens(self, n: int):
+        raise NotImplementedError
+        # for layer_idx in range(len(self.key_cache)):
+        #     head_num = len(self.metadata_list[layer_idx].head_lens)
+        #     cache_idx = torch.arange(0, self.key_cache[layer_idx].shape[0] - n * head_num, dtype=torch.int32, device=self.key_cache[layer_idx].device)
+        #     pass 
+
+        #     self.metadata_list[layer_idx]._update_metadata_remove_n(n)
+        # self._seen_tokens -= n
 
     def get_max_length(self) -> Optional[int]:
         return None
